Refactor out folder / url tabs into separate components from resource picker popover
